chest_xray/webapp/requirements.txt
# Web App - Chẩn đoán Viêm phổi từ X-quang
# ============================================

# FastAPI & Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.2

# ML & AI
tensorflow>=2.15.0
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
scikit-image>=0.21.0
